third_party/WebKit/LayoutTests/crypto/digest-failures-expected.txt

CONSOLE ERROR: Invalid dataBuffer argument
CONSOLE ERROR: Invalid dataBuffer argument
CONSOLE ERROR: Algorithm: Not an object
CONSOLE ERROR: Algorithm: Unrecognized algorithm name
CONSOLE ERROR: Algorithm: name: Missing or not a string
Tests incorrect calls to crypto.subtle.digest()

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS crypto.subtle.digest({name: 'sha-1'}) threw exception TypeError: Failed to execute 'digest' on 'SubtleCrypto': 2 arguments required, but only 1 present..
PASS error is null
PASS error is null
PASS error is null
PASS error is null
PASS error is null
PASS successfullyParsed is true

TEST COMPLETE
